+++ Trans3_2/vc/tkCommon/board/coords.cpp
@@ -88,8 +88,8 @@ void coords::pixelToTile(int &x, int &y, const COORD_TYPE coord, const bool bRem
 	{
 		case TILE_NORMAL:
 		{
-			x = floor(x / 32.0) + 1;
-			y = floor(y / 32.0) + 1;
+			x = int(floor(x / 32.0) + 1);
+			y = int(floor(y / 32.0) + 1);
 		} break;
 		case ISO_STACKED:
 		{
